Poll fsdb server health in subscribers
Summary:
In the event that fboss host userver panicking or otherwise shutting down ungracefully, fsdb subscribers can dangle forever and not reconnect. This is because the kernel does not have time to send a FIN on the tcp socket so the socket stays open.
Adding a change here to periodically poll fsdb health and if the server's alive since is newer than the last time a chunk was recieved we determine that a restart must have happened so we can teardown the stream and reconnect.
Technically since our retry interval is so low, it may be enough to just poll fsdb health as it will take longer for fsdb to come back up, but this seems more robust to me.
I'll remove this in favor of a proper heartbeat once that is available

Reviewed By: jasmeetbagga

Differential Revision: D45322300

fbshipit-source-id: 439ea0a993d2e9f3c2e882be9e66fcbbcbd36459