Revert D45322300: Multisect successfully blamed D45322300 for test or build failures
Summary:
This diff is reverting D45322300
D45322300: [fsdb] Poll fsdb server health in subscribers by peygar has been identified to be causing the following test or build failures:

Tests affected:
- [centralium/agent/tests:fsdb_subscriber_test - FsdbSubscriberTestFixture.StreamStateRemoveEntry](https://www.internalfb.com/intern/test/844425014839510/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2177779
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: cenzhao

Differential Revision: D46418526

fbshipit-source-id: 9f7feec0978aa64b8d6a296ba6d579f278fb73b0