Send Packets Directly to Data Plane for Load Balancing Invariant Test
Summary:
Instead of sending packets to a downlink port, I send packets to the data plane, which will forward these packets to a port in the ECMP group. Afterwards, based on hashing, the packets will appear in the counter corresponding to the appropriate ECMP port.

I also leverage the WITH_RETRIES and EXPECT_EVENTUALLY_TRUE utilities to run the load balancing check with retries and sleep in between to account for devices that take longer to update port stats.

These changes should stabilize ProdInvariantTests for wedge400 devices on both test and prod config runs.

Reviewed By: shri-khare

Differential Revision:
D46465494

Privacy Context Container: L1125642

fbshipit-source-id: 8d4fbeaebfc5273760b1e96d77ef9a4d04652b08