Rename FruConfig to FruTypeConfig in config
Summary:
This provides a clear way to differentiate FRU names provided in the
config from the names we assign to FRUs once they are discovered. For example,
a FruTypeConfig will be defined for PIM-8DD type of FRU.  If two such FRUs are
discovered in the platform, they might be assigned the names -
"Chassis/PIMSlot0:PIM-8DD" and "Chassis/PIMSlot1:PIM-8DD". In the code,
"PIM-8DD" will be referred as FruTypeName, and "Chassis/PIMSlot0:PIM-8DD" and
"Chassis/PIMSlot1:PIM-8DD" will be referred as FruNames. The FruNames are the
ones which will be exported in the platform_snapshot. They will also be the
unique identifiers used for listing FRUs in the platform manager interface.

Reviewed By: tfangit

Differential Revision:
D46471211

Privacy Context Container: L1125642

fbshipit-source-id: 474c45c9b618003a8b42e8d0e9ffd9d8753b2298