Correct clearing of error bits in DOM FPGA
Summary:
In Minipack/Fuji platforms, before every clause45 read/write operation, we write '1' to clear the mdio error and done bits. After this we immediately check if these bits are cleared or not.
First of all, we should give some time to FPGA to clear off the bit instead of checking the cleared bit once again.
Secondly, we don't need to wait for actual clearing of the error bits because the FPGA may take some time to clear off the stuck MDIO transaction. If there is a real error then subsequent read/write will anyway fail;. So this extra check is redundant and we need not crash out from here.

Reviewed By: harshitgulati18

Differential Revision: D44442133

fbshipit-source-id: b6be7409925fb97423ae51b2ae695e26a031af5d