Find the right queue to modify the config for in QosUtils
Summary:
As part of AqmTests, ECN/WRED queue configs are set/modified, however,
the current logic indexes based on queueId. Instead, we should look for the
queue cfg with specific ID before modifying the queue config to make sure it
works in all cases as the index of queue cfg is not really queue ID.

This is towards enabling ECN/WRED tests for VoQ platforms, running into issues
due to the fact that queues used in VoQ systems has different queue IDs as
compared to the ones used in non-VoQ and breaks due to the current assumption
about queue ID being the index.

Reviewed By: srikrishnagopu

Differential Revision: D46711850

fbshipit-source-id: 1663272c7887796dd5c7cd0947fdee8828bdd09d