Add fabric reachability utils for detecting missing and mismatching scenarios.
Summary:
Mismatch  -> If the attributes expected on neighbor (based on cfg) mismatch with one obtained from the neighbor reachability protocol.  This points to wrong or no connectivity scenario. This is cabling issues.

Missing  -> if the attributes expected on neighbor  are missing from the cfg, port is missing from cfg or the relevant  DSF node information is missing etc. as a result the neighbor reachability protocol won't be able to populate all fields.  This is mostly configuration issues.

Reviewed By: zechengh09

Differential Revision:
D46787202

Privacy Context Container: L1125642

fbshipit-source-id: bc015d452dd338e0f60826465d37893166809377