Revert D46292879: Multisect successfully blamed D46292879 for test or build failures
Summary:
This diff is reverting D46292879
Depends on D46823329
D46292879: [fboss] apply thrift config to compute and inject seeds to load balancer config applier by phshaikh has been identified to be causing the following test or build failures:

Tests affected:
- [fboss/agent/hw/bcm/tests:fake_bcm_test-6.5.21 - HwAclStatTest.AclStatRename](https://www.internalfb.com/intern/test/844425009961437/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2310905
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: srikrishnagopu

Differential Revision: D46823335

fbshipit-source-id: cc006c66365e7cbb5bc7774afa49b1009f42e62a