use platform mapping passed in apply thrift config
Summary:
instead of using platform to get platform mapping, use the platform mapping passed as an argument to apply thrift config.

this is a straight forward change.

however this breaks override mechanism for inter packet gap bits added in D31419312

this was added in BCM test platform (and test platforms only), to make the test to verify inter packet gap bits work on all platforms. this was working as apply thrift config used platform to query platform mapping and not platform mapping passed as an argument.

override mechanism was available only in platform and not in platform mapping. there is no way to modify platform mapping (as of yet) and didn't see the need to add that api for a single test

Reviewed By: zechengh09

Differential Revision:
D46903040

Privacy Context Container: L1125642

fbshipit-source-id: 302deaa30dd2da19f34a30253194c2aa21063bc0