Clear led state when lldp cabling error goes away
Summary: D46889343 moved lldp cabling error to switch state. It followed the same logic as old code where on detecting cabling error, the led is set but there is no explict clear if the error goes away. It is better to do an explict clear as switch state is persistent across reloads

Reviewed By: rajank7

Differential Revision: D46960350

fbshipit-source-id: e121232bba19e732968a798cea4e4dafd63a0eb1