Heartbeats(3/n) Handle heartbeat special case for OperState
Summary:
In the current chunk structs heartbeats will be encoded as empty chunks, which will be ignored by the client library. WIth OperDelta we have a list of changes and an empty list is already ignored in client libraries. With OperState unfortunately we don't have that luxury, an empty contents will be interpreted as a deletion at that path. So for now adding a flag to indicate heartbeat and ignoring accordingly. After this is pushed to subscribers we're safe to enable heartbeats everywhere

Long term these structs will be replaced and we can keep heartbeats in mind when designing new ones

Differential Revision:
D47026580

Privacy Context Container: L1125642

fbshipit-source-id: 312be6858f25c4a83b9d86d775a375aae3638a15