Heartbeats(5/n) Add options to set chunk timeout in stream clients
Summary:
Adding option to enforce a chunk timeout. With this the client disconnects when a chunk is not recved in the set time. This will work with the heartbeat to make sure when the server disappears for any reason the client will know.

Introduced two gflags since the heartbeat interval for state/stats is different. For now both are set to 0 because it's not safe to enable this enforcement until we know the server is serving heartbeats (else we'll see repeated disconnects). We'll need to carefully enable this for subscribers

Differential Revision:
D47069788

Privacy Context Container: L1125642

fbshipit-source-id: e16eb11deecac6e6f18eb69cf5877c8ddb77a129