Add credit watchdog delete counter to HwSysPortStats
Summary:
D47179475 mistakenly added this to HwPortStats. Since
this counter only applies to VOQs, add this to sys port stats.
Subsequent diff will prune this from HwPortStats.

While at it also fix the half done wred stat for VOQs. Note
that the stat is lazily populated, so we will only set it
if the feature is enabled and stat is read. So in prod, we
won't really burn counters for this.

Reviewed By: shri-khare

Differential Revision:
D47199494

Privacy Context Container: L1125642

fbshipit-source-id: fd4b170c51e504d6bf8e4eadf7805d10b32495bc