Move AccessMethod parsing to thrift
Summary:
The AccessMethod thrift struct was defined using thrift, but was
parsed using folly::dynamic and json. This diff fixes it. One downside of this
change is that the config will have enum values (ints) for the accessType
instead of string, reducing readability. This can be changed to string as well,
but is a more involved change.  It will be easier to make that change after
rest of the fan_config_structs has been moved to thrift serialization

Reviewed By: mikechoifb

Differential Revision: D47266821

fbshipit-source-id: c26eb28301913e248d9b42cb7e39982cf13a65b1