SwitchSettings fixup for missing fields
Summary: Some fields in switch state were moved to switch settings in D43968151 back in march timeframe. Later the switch settings struct in switch state was moved to SwitchSettingsMap in D45004833 in april. Each of these transitions had logic to write the fields to old and new places. The first transition was rolled out to fleet as part of agent version 630. There were 180 rtsws where push skipped upgrade to 630 version due to a broadcom issue with fhte. On these switches, the transition logic for switchSettings to switchSettingMap would run first before the transition logic for switchstate to switchSettings field migration resulting in these fields being empty in the map. Recently in version 635, we stopped using fallback to original fields which causes warmboot failures on devices which skipped 630

Reviewed By: aeckert

Differential Revision: D47356473

fbshipit-source-id: af9e7bf11b49d3a7e4b1941e3077321a939469b7