Semi-deprecated nextHops field and replaced it with an optional nexthops field
Summary:
# What?

This diff introduces a new optional field of `nexthops`, which will be used going forward for the programming of the TE flows.

NOTE: During migration, both of these fields will be populated by TE agent.

# Why?

This is needed for implementing the new "inactive" exact match (EM) entries.

The new inactive EM entries will be used to solely monitor/count the traffic flows, instead of to steer traffic like the "active" (original/existing) version.

The proposed new contract for `nexthops`:
- nullopt: no nexthop action ("inactive" EM entry)
- empty list: null route
- non-empty list: nexthops

This new contract is consistent with the contract of the nexthops field for other routing protocols.

Please see the design/alignment doc https://fb.quip.com/ymLBAAIhscuI and also the task T155874018 for more details.

Reviewed By: msomasundaran

Differential Revision: D47232245

fbshipit-source-id: 2cdb80d58ef0facd1fe5ff35302fcedf9743519e