Refactor fields used in sensor pwm calculation
Summary:
In the current code, there are three cpp classes - `IncrementPid`, `FourCurves`
and `Sensor`. The config values used in pwm calculation as well as the
intermediate state of pwm calculation are stored in `IncrementPid` and
`FourCurves`.  `Sensor` class only has half of the config values for the
sensor.

This diff changes the above to the following. `Sensor` class has all the fields
set in the config for the sensor. `IncrementPid` and `FourCurves` are folded in
`PwmCalcCache` class, and it contains only the intermediate state used in pwm
calculation.

The purpose of this change is to replace `Sensor` class with `Sensor` thrift
struct in the subsequent diff. The `Sensor` thrift struct should hold all the
values specified in config for the sensor and only those values.

Eventually, the pwm calculation can be modularized into testable entities
(later).

Reviewed By: mikechoifb

Differential Revision: D47419927

fbshipit-source-id: 7e41a069d38a137dd170172ee56aa5406fc54f4d