Move intermediate state holders to ControlLogic
Summary:
The three classes - `FanStatus`, `SensorReadCache` and `PwmCalcCache`
were used for holding intermediate state for pwm calculation. This diff makes a
couple of changes.

- Changes the above three to structs
- Moves them to ControlLogic module from ServiceConfig

The above change makes sense these three structs are exclusively used only by
ControlLogic class. They are local state for ControlLogic class.

After this change ServiceConfig is entirely read-only, and it is the user
provided config. This in turn facilitates complete replacement of ServiceConfig
with a thrift config in subsequent diff

Reviewed By: mikechoifb

Differential Revision: D47444730

fbshipit-source-id: 16a6dd29ac4303d12ce132cb7ff7ac001effc9ba