Refactor getControlFrequency and getSensorReadFrequency
Summary:
`FanService` had a getControlFrequency method which returned a value
which is never set in the program.  `FanService` though used
`ServiceConfig::getControlFrequency` for all its logic.
`ServiceConfig::getControlFrequency` returned the intended value, so everything
continued to work as is.

In this diff, we remove the `getControlFrequency` and `getSensorReadFrequency`
methods from ServiceConfig. Instead we make the same methods in `FanService` as
the source of truth.  This addresses the issue stated above, as well as ensures
`ServiceConfig` has only user provided config parameters for the platform, and
not the execution parameters of the binary.  Eventually, we need to move these
two to be controlled by flags.

Reviewed By: mikechoifb

Differential Revision: D47448397

fbshipit-source-id: 42868667e33027fa3e189ab512edb7423e3b19cc