Handle all port types in default lb mode map.
Summary:
For J3 support we added a new lb mode for
Interface ports. This required us to build a port type
to LB mode map. This is used as a optional param for
config generation. However the default map value does
not handle all port types, causing failures on DNX devices

Reviewed By: srikrishnagopu, zechengh09

Differential Revision: D47522929

fbshipit-source-id: 05e17a42e5096bc00dbe38bc0a8429be6734286f