Link Test to verify fabric link health
Summary: This test utilizes the fabric port spray feature. It pumps traffic out of the cabled NIF ports. Because the ports are snaked, packets come back to the snaked nif port and are then sprayed on the fabric. The test intends to verify the health of the fabric links. It currently verifies the fabric links receive cells. Will add more checks to this test in the coming diffs. Details are in the comments.

Reviewed By: jasmeetbagga

Differential Revision: D47425253

fbshipit-source-id: e6591413b1c2ea6fcab7c85a316787bf68d75d36