Model Switch Main Board as a FRU
Summary:
In the current code the Chassis is modeled as a FRU, with the chassis
FRU being plugged into BSP/main_board.  Chassis FRU is more of a logical
contruct, since Chassis is the housing for the entire switch. This diff
proposes the main-board as FRU which gets plugged into the SCM (switch control
module) / CPU. This maps to the physical layout in the switch. The SMB gets
plugged into SCM.

I have using the following naming
- MainBoard for SMB
- CPU for SCM.

I am not entirely sure about the naming. Happy to adopt SMB/SCM naming if it is
a more generic/relatable name.

The new modeling in sample.json is captured in
{P770900003}

For context, the old modeling is
{P714913422}

Reviewed By: joancaneus

Differential Revision: D46746868

fbshipit-source-id: a6438cbc11d374c86c94d5962cbf5054a9911bf8