Group configs by platform type
Summary:
This diff does two things

## Changes the directory structure of configs from
```
fboss/platform/configs/
├── darwin
├── fan_service
│   └── darwin.json
├── fbdevd
│   └── darwin.json
├── fw_util
│   └── darwin.json
├── platform_manager
│   └── sample.json
├── sensor_service
│   └── darwin.json
└── weutil
    ├── darwin.json
        └── montblanc.json
```

to
```
fboss/platform/configs/
├── darwin
│   ├── fan_service.json
│   ├── fbdevd.json
│   ├── fw_util.json
│   ├── sensor_service.json
│   └── weutil.json
├── montblanc
│   └── weutil.json
└── sample
    └── platform_manager.json
```

This ensures the configs are grouped by platform and not by service type.

## Changes the config generation logic accordingly.

The internal state still groups configs by serviceName (see GeneratedConfig.h
in paste). But this is an internal state and can be changed later if needed.
This should not be visible in any external system.

Reviewed By: tao-ren

Differential Revision:
D47573633

Privacy Context Container: L1125642

fbshipit-source-id: 9161e7bae009b68bcf404f1fabc02740c5b7497e