Fix for FSDB SwitchState info to send to LedManager
Summary:
In LED Service, the state subscriber callback function was earlier sending the SwitchState.PortMap. This is complex structure with multiple structures embedded. The subscriber has to create a copy and send to future function of LedManager thread. Since the structure is complex, the deep copy operation used to fail and resulted in crash.

In this diff we are creating a map of relevant SwitchState info which we need to send to LedService and sending this data from Subscriber callback to the LedManager thread.

Reviewed By: harshitgulati18

Differential Revision:
D47736075

Privacy Context Container: L1125642

fbshipit-source-id: 305509e957b45f3b17446eb6f9c06917bc9cdb67