Add defaultVoqConfig to switch_state
Summary:
As titled. Also, introduce apis to get and set defaultVoqConfig in switch setting.

The purpose of adding this in switch state is to be able to get delta processing done if this
config changes for VoQ switches. The defaultVoqConfig is a device level config, however,
this gets copied to the port state for each port so that it gets applied per port during programming.

However, if the config changes, we need to have a trigger which can let the resetSystemPorts
flow get hit, which will be based on the switch settings delta as done in the next diff.

   OK : 3
   FAILED : 0
   SKIPPED : 0
   TIMEOUT : 0

Test output stored at: hwtest_results_2023_Jul_24-11_34_17_AM.csv
[root@meru400biu.13.09.mpk17.fbinfra.net ~]#
```

Differential Revision: D47612516

fbshipit-source-id: 550b4d468711aff15bfc8d22be496d05e0dec249