hw switch route updater refactored
Summary:
refactoring hw switch route updater to fit hw switch ensemble route updater in the hierarchy.

route updater is fitted in such a way that a method to "apply state update" can be injected.

note that this method is not necessarily hwSwitch->stateChanged(delta), because the invocation of applying state update may be interwoven in different places of code such as when invoking stateChanged method is not possible (such as during cold boot) or in ensemble where state updates converge into 'apply new state' method.

as a result an 'apply' argument is added and typically common fibUpdateFunction is provided by default. fibUpdater updates a state and invokes apply on stateDelta

this will be used in hw switch ensemble route updater

Reviewed By: jasmeetbagga

Differential Revision:
D47748063

Privacy Context Container: L1125642

fbshipit-source-id: 90dce00e17d0994c7ba5d808a64b129a70412bac