Add IP Address Check for Switch Config Port Interfaces
Summary:
On switch configs of Galaxy devices, some interfaces that ports are being linked to are being unused. As a result, when such ports are added to the ECMP group, invariant testing crashes because test setup is unable to compute ECMP next hops for these ports.

To resolve the issue, I implemented a check to ensure that all uplink/downlink ports being used in invariant testing are linked to an interface that contains at least one IP address.

This fix should stabilize ProdInvariantTests for Galaxy linecard devices.

Reviewed By: zechengh09

Differential Revision:
D47848984

Privacy Context Container: L1125642

fbshipit-source-id: 0e81b443681cd48cd63fc6396732e80909d6a62a