Work towards supporting default values for ext attributes
Summary:
We need this mechanism for extension attributes as well.
Case in point, we have extension attributes for VOQ switch that
don't apply to fabric switches. So we need to allow fallback
to default values there or build conditional attribute classes
for all such apis. Former is better and inline with existing
SaiAttribute mechanism

Reviewed By: shri-khare, zechengh09

Differential Revision:
D48080857

Privacy Context Container: L1125642

fbshipit-source-id: 7af481423aa4ca91074709f3bd04755371e21ddf