Add patch version to identify incremental SDK releases for TAJO
Summary:
Currently SDK releases are of the format 1.42.1, 1.42.8 etc. When there are bug fixes on top
of these releases, a new patch release is made which is of the format 1.42.8.2, where 2
identifies the patch release. Making changes to show the patch version along with major
version in the show version o/p for SDK.

Reviewed By: srikrishnagopu

Differential Revision: D47886073

fbshipit-source-id: 920e08ed93d17161d90d774797614b803441e8a7