set warmBootStateWritten_ in HwSwitchWarmBootHelper
Summary:
setting this to true after saving warm boot state file.

this is used in BcmSwitch to touch warm boot flag file

this is different from how it is done in SaiSwitch, and a clean up is needed to unify how warm boot helper is used in both SAI and BCM switch.

Reviewed By: msomasundaran

Differential Revision:
D48208539

Privacy Context Container: L1125642

fbshipit-source-id: eb2a14f1d94f39bc75be83ba5581b069cb52cd3a