simplifying hw switch warm boot helper
Summary:
removing a boolean member variable warmBootStateWritten_

this variable was used in BcmSwitch to determine if the warm boot state is dumped or not. However not able to dump warm boot state is fatal error and should lead to hard crash. As a result this flag is unnecessary. This flag was later tested in BcmUnit to determine if warm boot flag needs to be set up or not (warm boot flag file is touched or not). If warm boot state is dumped, set this flag right away, removing another call to touch warm boot flag file. made setCanWarmBoot file private

also removed ununsed method

Reviewed By: msomasundaran

Differential Revision:
D48215493

Privacy Context Container: L1125642

fbshipit-source-id: 1707f930fe272b7ff3ef7e51a842210eea627fbb