Release previously configured lanes in a datapath before a new configuration
Summary:
We need to release previously configured lanes in a datapath before applying a new configuration. For example, on a 2x400G-FR4 module the default configuration is 400G on lanes 0,1,2,3. If we then need to program 100G-FR1 on lane 2, we should first release (apply app sel 0) to lanes 0,1,2,3 and then apply 100G application on lane 2.
The following diff handles that.

Reviewed By: rajank7

Differential Revision:
D48142121

Privacy Context Container: L1125642

fbshipit-source-id: 9ca84329dd33c2472a2aed6c5b9ff9f47ff65664