rackmon2: Fix use-after-scope-end during test runs
Summary:
This was caught by Address Sanitizer. Essentially we
have the `mon` object which uses the `commandTimeout`
bool. When the function scope is being terminated,
before this change commandTimeout is destroyed and then
mon. This causes a race where mon's threads could access
the now destroyed commandTimeout.

Imported from: D48085535

Reviewed By: somasun

Differential Revision: D48086183

fbshipit-source-id: ece24653786bc0c0d381dc7eec71569f0e7ec9f8