rackmon2: Start monitor immediately after discovering devices
Summary:
"tick" the monitor poll thread to immediately start reloading
registers as soon as we discover them. This should fix the issue
where we need to potentially wait like 3min after scan for
the monitored data.

Now that the "poll threads" can be accessed from outside their thread
context, it becomes unsafe to access the pointers without having a
reference. Even if we check for validity of the pointer, it cannot be
trusted that some other thread has requested for the thread to stop
and hence deleting the reference. Hence convert them to shared_ptr
that way they are reference counted.

Imported from: D44385226

Reviewed By: somasun

Differential Revision:
D48053374

Privacy Context Container: L1125642

fbshipit-source-id: dca48f4bff5f192f9896f5233fa96f0ce1aeaf58