save local node name in DsfSubscriber
Summary:
We need the local node name to refine our subscription to
remote subscription status in remote fsdb instances. This info is part
of the DsfNodes map and we need this information as soon as we process
any DsfNode delta in DsfSubscriber. Unfortunately, order is not
ensured and we learn our dsf nodes all at the same time. We Could try
to ensure we always process added nodes for ourselves first, but it
seems better to get our dsf node name through a more direct method.

For now, we will use the local hostname uqdn as the local node
name. I'm going to add a dedicated config field for localNodeName to
remove the call to get hostname.

Reviewed By: zechengh09

Differential Revision: D48195236

fbshipit-source-id: 5b437da7245f564767e1263f63409c615b8b0c2f