Model main_board fru as one of the fruTypeConfigs
Summary:
Previously main_board fru was modeled as a special fru, though all the
fields apply to it like any other fru. The only difference between the
main_board fru and other frus is that main_board is always expected to be
present. In this diff, we model fru as one of the fruTypeConfigs. In validation
code we enforce that main_board FruTypeConfig is always present.

The advantage is simpler code in PlatformExplorer. We can explore MAIN_BOARD
SlotConfig as any other SlotConfig

Reviewed By: tfangit

Differential Revision: D48197842

fbshipit-source-id: 72840a1b96c97cc8cd56605e309edf14624cc283