Use dmidecode to determine platform name
Summary:
If a --config_file is provided, that will take priority. If there is error in
parsing this config_file, then the program exits.

Else, the code will read dmidecode to get the platform name and load the
corresponding config. If no config can be fetched for the inferred platform
name, then the program exits.  If there is failure to read dmidecode, then the
program exits.

Differential Revision: D48073429

fbshipit-source-id: fa41ca04ef0d908a5738b3a3e44eeda2a9b219a2