Don't need to unisolate in HwLinkStateDependentTest::Setup
Summary:
For VOQ switches:
 - switch is created with isolate = true
 - ApplyThriftConfig computes actualSwitchDrainState
 - actualSwitchDrainState = desiredSwitchDrainState if min VOQ thresholds are not set
 - The default desired drain state is UNDRAINED
 - Thus, the default (UNDRAINED) gets applied and brings the switch out of isolate

As a result, we no longer need to explicitly unisolate the device in
HwLinkStateDependentTest::Setup.

Note: stacked diffs will explicitly set min VOQ thresholds for VOQ switches -
which will leave the device in isolated state - but there we will also add
logic for the linkscan callback processing to schedule state update to bring
the device to unisolated state once the threshold is hit.

Reviewed By: jasmeetbagga

Differential Revision: D48383482

fbshipit-source-id: 724888376ec79e1cdecc6532bf79def23c06cb87