Outline of HwTestLinkScanUpdateObserver
Summary:
As detailed earlier in this stack, VOQ switches need to distinguish between
desired vs. actual switch drain state.

Actual switch drain state is computed dynamically based on the number of fabric
links that are up + pre-configured threshold.

In order to write a HwTest to verify this fucntionality, we need HwTest to be
able to schedule state updates on linkscan events. Thus, start adding an
observer.

Stacked diffs will leverage it.

Reviewed By: jasmeetbagga

Differential Revision: D48379889

fbshipit-source-id: 5e40a6a4a6926282106e843f97dce2c0970257ed