Skip DHCP processing for VOQ switches for now
Summary:
VOQ/Fabric switches require that the packets are not tagged with VLAN.
We are gradually enhancing the wedge_agent to handle tagged as well as untagged
packets.

DHCP handling code relies on presence of VLANs today. Thus, skip DHCP
processing for VOQ switches till that code is enhanced to work with Interfaces
only.

Reviewed By: jasmeetbagga

Differential Revision:
D48497504

Privacy Context Container: L1125642

fbshipit-source-id: c35cf9e6b50d78535382a0a3bb84268d987579cc