InterfaceTests: remove pseudo vlan support
Summary:
VOQ/Fabric switches require that the packets are not tagged with VLAN.
This diff is part of the diff stack that enhances wedge_agent to deal with VLAN
tagged as well as untagged packets.

We introduced a notion of pseudo VLANs (vlanID = 0), to allow us to gradually
convert agent code to work with no-VLANs. In particular, NeighborUpdater was
not updated to work without VLANs then.

That is fixed since, and VOQ switches always run with intf_nbr_tables
flags enabled. Thus, remove pseudo vlan.

Reviewed By: jasmeetbagga

Differential Revision:
D48498920

Privacy Context Container: L1125642

fbshipit-source-id: 98a86f8d15830ed50006f81e941330935447d257