InterfaceTests: remove pseudo vlan tests
Summary:
VOQ/Fabric switches require that the packets are not tagged with VLAN.
This diff is part of the diff stack that enhances wedge_agent to deal with VLAN
tagged as well as untagged packets.

We introduced a notion of pseudo VLANs (vlanID = 0), to allow us to gradually
convert agent code to work with no-VLANs. In particular, NeighborUpdater was
not updated to work without VLANs then.

That is fixed since, and VOQ switches always run with intf_nbr_tables
flags enabled. Thus, remove pseudo vlan unit tests.

Stacked diffs will remove the Agent pseudo vlan code as well.

Reviewed By: jasmeetbagga

Differential Revision: D48498919

fbshipit-source-id: c6244109d6d5547815397ce443ffc0cfc9ce6946