SwSwitch to read warm boot state and seed to HwSwitch
Summary:
this stack introduces following changes
1) on warm boot, SwSwitch (and HwSwitchEnsemble), reads switch state and seeds to HwSwitch init
2) MonolithicHwSwitchInitFn is a function which is invoked from monolithic agent which does both init and initial warm boot state apply.

Note there is a hack in this diff which is rectified subsequently. HwSwitchWarmBootHelper uses SwSwitchWarmBootHelper and so the warm boot flags are cleared during initSaiPlatform. To fix this, a method is made public to restore flags of SwSwitchWarmBootHelper.

This is being fixed in subsequent diffs.

Reviewed By: jasmeetbagga

Differential Revision: D48177629

fbshipit-source-id: bf78c706e8d5fd441aa44dca9f739e7b6e90a290