Disable remediation for meru400biu and meru400bfu
Summary: On these platforms, we are configuring the 200G optics in 2x50G experimental mode. Thus 2 of the 4 lanes remain disabled which kicks in the remediation logic and flaps the other 2 ports. Disabling remediation for just these 2 platforms as this is an experimental mode only.

Reviewed By: rajank7

Differential Revision:
D48550041

Privacy Context Container: L1125642

fbshipit-source-id: b7d6472dd828688dc2307be27bf833602543f3d4