Ensure VOQ DELETE counter gets called if feature enabled
Summary:
Rename supportedCounterIds to basePlusWredCounterIds to make the purpose explicit.

Also, Fix the sequencing to include VoqWatchDogDeleteCounterId to the voq counters to
be  collected if feature is enabled by first adding to the base counter IDs. This way,
VoqWatchDogDeleteCounterId will be polled with or w/o WRED being enabled on queue.

Reviewed By: jasmeetbagga

Differential Revision: D48182058

fbshipit-source-id: 10fe41f477b5e9c0655075fb158bf2fc960cdd14