Mechanism for syncing oss thriftpath
Summary:
thrifpath is generated using a buck thrift plugin. These are not supported in oss so to opensource thriftpath, I've opted to copy over and commit the generated header for fsdb model. With this, FsdbModel.h will pick the generated header if building with buck but use the synced file when building in oss. For I'm syncing the file once, but I will add some process to keep this file in sync.

fsdb_model_thriftpath.h is the auto generated file which the script synces

To streamline the process, I think I will start with adding an error on diff when the model is changing and tell engineers to run the script before commiting (this can be done with a custom unittest). This will trigger anytime an fboss engineer commits a change to our thrift model. If needed we can also add an automated chronos job to sync.

Differential Revision: D48200593

fbshipit-source-id: b14cfdd899cbb244f71ac346c4112a13f6468e0c