Added absolute reference to facebook::stats::MonotonicCounter
Summary:
MonotonicCounter.h defines facebook::stats namespace
There is qsfp_stats_types.h that defines the facebook::fboss::stats namespace

In the next diff, there is a file that transitively includes both MonotonicCounter.h and qsfp_stats_types.h. When this happens and without the change in this diff, the compiler looks for MonotonicCounter in facebook::fboss::stats namespace since that exists when both the offending files are present. The compiler doesn't find MonotonicCounter in facebook::fboss::stats namespace and errors out.

Fixing it here by changing it to the absolute reference to MonotonicCounter in it's appropriate namespace.

Reviewed By: shri-khare

Differential Revision: D48648240

fbshipit-source-id: eefd2fabb356413a6e95c8bb12901b59f4075bf5