Skeleton for SplitAgentTest
Summary:
As suggested in comment -
SplitAgentTest is the base class for agent tests running in mono or split mode. At early phases, the tests will be running in both mono and split mode to assist migration. This should eventually be named as AgentTest as all Agent/Hw tests are migrated to this framework.

Reviewed By: jasmeetbagga

Differential Revision:
D48700765

Privacy Context Container: L1125642

fbshipit-source-id: f2c7a0d7fb39d7dea90582149f721f2b8a5a5fa6