DHCPv6HandlerTest: remove now unused helper methods
Summary:
VOQ switches don't support VLANs.
Thus, we are in the process of migrating wedge_agent to use neighbor tables
from Interfaces.

As part of this effort, work towards getting DHCP relay working without VLANs
as well.

Previous diffs in the stack created and migrated to using helpers that are
templatized for vlans as well as intfs. Thus, these helpers are no longer used.
Remove them.

Reviewed By: jasmeetbagga

Differential Revision: D48849387

fbshipit-source-id: bdf29da051c64af235cd4e39e2aa51d526a812e8