Add utlity to add TTL0 cpu trap config.
Summary: In DSF, inroder for tests to create traffic loop, we need to set a rxreason (trap) which will continue to forward the traffic even if TTL = 0.  This function utility intends to configure TTL0 rxreason and since it applies only to DSF its gated by a feature.

Reviewed By: srikrishnagopu

Differential Revision:
D48860600

Privacy Context Container: L1125642

fbshipit-source-id: 2a62a94a3ecbbee6e0b50a0c03daffddbe8b2c33