Handle the list objects which needs to fallback to default.
Summary:
With the list type defaults,  if the nullopt is programmed and if SAI layer returns not supported/error, we  provide the default list parameter which results is not equivalent to null_opt.
As an example:

SaiStore set object HashSaiId(120259084289): (NativeHashFieldList: [0, 1], nullopt)

/* warm boot */
V0829 00:00:44.168564 4024331 SaiStore.h:403] HashSaiId(120259084289): (NativeHashFieldList: [0, 1], UDFGroupList: [])

This is a problem, since 2 objects are not equivalent.

Fix is to return null_opt during warm boot if the object has default value of type list/vector.

Reviewed By: jasmeetbagga

Differential Revision:
D48816828

Privacy Context Container: L1125642

fbshipit-source-id: ba5032fad7f19bbd3c4dc8a8d5be502470463b97