Added functions to hold and release reset individually
Summary:
In this diff stack, i want to add a qsfp hardware test to make sure reset lines work correctly on transceivers. Today, we have a triggerQsfpHardReset() function which does a combined trigger reset + clear reset logic.

In this diff, I separate out the hold and reset logic that will make it easier to test each aspect of this operation.

Currently implementing it for BSP platforms, we can extend it to other platforms later.

Reviewed By: rajank7

Differential Revision:
D48992571

Privacy Context Container: L1125642

fbshipit-source-id: 6928d5122cf9138c802899fc7cd3a2f21473027f