enable logging in OSS for wedge agent and hw test
Summary:
As titled, enable logging in open source by setting it to DBG4.

The filename provided to logger uses the string "fboss/agent/.cpp".vLoggerDB recursively finds the parent by stripping the filename usingvthe separator and finally ends with "fboss" in the string and thevcorresponding LogLevel is picked to print the log.

In OSS, the file name provided to logger uses the stringv"/var/FBOSS/tmp_build_dir/fboss-git/fboss/agent*.cpp". When LoggerDBvrecursively finds the parent by stripping the filename using the separator, it never finds "fboss" string in its DB and ends up setting the default logging level which is INFO.

The stripping is different internally and in OSS because when buck2 is used to build, there is an additional processing done to the filenames which removes few strings from the beginning of the path.

To fix this issue, we are adding a flag for OSS builds to set the logging without the category.

Reviewed By: shri-khare

Differential Revision: D49088582

fbshipit-source-id: 4a0d5705ee7278db8da3a0c136794151c93b7371