synchronize access to agent config in SwSwitch
Summary:
there is a race between applyConfig and getConfig. applyConfig is expected to happen once or in response to reloadConfig (thrift API), while getConfig happens in thrift APIs.

But there're tests such as

```
LookupClassRouteUpdaterTest/[01].CompeteClassIdUpdatesWithApplyConfig
```

where getConfig and applyConfig are invoked in separate threads which exposed this need for synchronizing access to agentConfig_

Reviewed By: jasmeetbagga

Differential Revision:
D49119983

Privacy Context Container: L1125642

fbshipit-source-id: a1414c5326265462ef6dc1e064c910b586097ac8