Skip programming ports with startHostLane > numHostLanes
Summary: It's possible that we plug in a module with lower number of lanes than the number of lanes supported on the transceiver slot. When that happens, we should skip program transceiver for the ports with higher lanes, otherwise we will remain in a constant loop of programming transceivers.

Reviewed By: rajank7

Differential Revision:
D49176014

Privacy Context Container: L1125642

fbshipit-source-id: afac5fcfc0a08fee5bafbf267f3b9b88b8c19b5a