Adding uppercase product name for Meru
Summary:
PlatformProductInfo library parses the fruid.json on the x86 to
determine the actual platform being used. The fruid.json is the content of the
eeprom which contains uppercase letter sometimes. Modifying the file below so
that it can work with both lowercase and uppercase platform name.

Reviewed By: somasun

Differential Revision:
D49262473

Privacy Context Container: L1125642

fbshipit-source-id: f4ef736420a0e8e4261ad7105629f50929c14688