Fix typo in Actual switch drain state computation
Summary:
This was introduced in D48098400, and see comments in the code viz.:
https://fburl.com/code/jfrpbs9i for why we need the notion of desired vs.
actual switch drain state.

The first version of the diff used names 'desiredSwitchDrainState' and
'switchDrainState' to mean actual.

We later decided to use the terminology 'switchDrainState' to mean desired and
'actualSwitchDrainState" to mean actual.

But one location was left unchanged. Fix it.

Reviewed By: jasmeetbagga

Differential Revision:
D49256354

Privacy Context Container: L1125642

fbshipit-source-id: 418850d4855bdbde76c35de62122096242ce754c