Added a UPGRADE_FIRMWARE event
Summary:
This event for now is only possible from the INACTIVE stage as we have decided to trigger firmware upgrade when the link goes down. When this event is triggered, we'll transition to UPGRADING state. As soon as we enter the UPGRADING state, we'll trigger the firmware download.

After Upgrade is done, we will reset the transceiver and go back to not present state, hence added a transition for UPGRADING to NOT_PRESENT state also.

Reviewed By: rajank7

Differential Revision:
D49303887

Privacy Context Container: L1125642

fbshipit-source-id: 3f065409d98ece16957f9f1e6673a78e3f003669