introduce sdk version to hw asic
Summary:
As titled, introduce  sdk version to hw asic.

HwAsic is built independent of the preprocessing flags that is set from the SDK. For Tajo, the default stream type need to be different for two different SDK versions. Hence leverage the SDK version at runtime and set the default stream type.

Platform has a hw asic instance which has the right default stream type set since platform is aware of the PP flags.

HwAsicTable which resides in SwSwitch creates an hw asic instance but SwSwitch is not aware of the SDK or the PP flags. Hence, in prod, we will be parsing the config, identify the SDK version (https://fburl.com/code/6bzm5pva) and then set the stream type based on the SDK version.

SdkVersion has two attributes - asic sdk version and sai version.
Only asic sdk version is set for Tajo whereas asic sdk version is set to native bcm sdK version and sai version set to bcm-sai sdk version for BCM.

Read the detailed summary in D49358544 for more details.

Reviewed By: msomasundaran

Differential Revision: D49358548

fbshipit-source-id: ad0b4636b3d0f592c55973ee8beba7df3fa27249